Added creating projects

+++ ToDo/Controllers/ProjectsController.cs
@@ -10,6 +10,7 @@ using ToDo.Infrastructure.Data.Models;
 using Azure.Core;
 using System.Text;
 using System.Linq;
+using Microsoft.VisualStudio.Web.CodeGenerators.Mvc.Templates.Blazor;
 
 namespace ToDo.Controllers
 {
@@ -34,13 +35,33 @@ namespace ToDo.Controllers
         [HttpGet("/projects/create")]
         public async Task<IActionResult> Create()
         {
-            await _projectService.CreateProject(User.FindFirstValue(ClaimTypes.NameIdentifier));
+            return View("Details", new ProjectDetailsVM());
+        }
+        [Authorize]
+        [HttpPost("projects/create")]
+        public async Task<IActionResult> Create([FromForm] ProjectDetailsVM details)
+        {
+            details.Participants = Request.Form["Participants"].ToString()
+                                                               .Split(",", StringSplitOptions.RemoveEmptyEntries)
+                                                               .Where(id => id != User.FindFirstValue(ClaimTypes.NameIdentifier))
+                                                               .Select(id => new ListedUser(id, ""))
+                                                               .ToList();
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
+
+            await _projectService.CreateProject(User.FindFirstValue(ClaimTypes.NameIdentifier), details);
             return View("Index", new ProjectsVM(User.Identity.Name, ProjectsTab.Created, await _projectService.GetCreatedProjects(User.FindFirstValue(ClaimTypes.NameIdentifier))));
         }
         [Authorize]
         [HttpGet("/projects/delete{id}")]
         public async Task<IActionResult> Delete(string id)
         {
+            if (!(await _projectService.ProjectExists(id)))
+            {
+                return NotFound();
+            }
             await _projectService.RemoveProject(id);
             return View("Index", new ProjectsVM(User.Identity.Name, ProjectsTab.Created, await _projectService.GetCreatedProjects(User.FindFirstValue(ClaimTypes.NameIdentifier))));
         }
@@ -48,19 +69,29 @@ namespace ToDo.Controllers
         [HttpGet("project-details/{id}")]
         public async Task<IActionResult> Details(string id)
         {
-            ProjectDetailsVMWithId projectDetails = await _projectService.GetProjectDetails(id, User.FindFirstValue(ClaimTypes.NameIdentifier));
+            ProjectDetailsVM projectDetails = await _projectService.GetProjectDetails(id, User.FindFirstValue(ClaimTypes.NameIdentifier));
+            projectDetails.Id = id;
             if(projectDetails.IsCreator) return View(projectDetails);
             return View("DetailsReadOnly", projectDetails);
         }
         [Authorize]
         [HttpPost("/projects/update")]
-        public async Task<IActionResult> Update([FromForm] ProjectDetailsVMWithId details)
+        public async Task<IActionResult> Update([FromForm] ProjectDetailsVM details)
         {
-            details.Participants = Request.Form["Participants"].ToString().Split(",", StringSplitOptions.RemoveEmptyEntries).Where(id => id != User.FindFirstValue(ClaimTypes.NameIdentifier)).Select(id => new ListedUser(id, "")).ToList();
+            details.Participants = Request.Form["Participants"].ToString()
+                                                               .Split(",", StringSplitOptions.RemoveEmptyEntries)
+                                                               .Where(id => id != User.FindFirstValue(ClaimTypes.NameIdentifier))
+                                                               .Select(id => new ListedUser(id, ""))
+                                                               .ToList();
+
+            if (!(await _projectService.ProjectExists(details.Id)))
+            {
+                return NotFound();
+            }
 
             if (!ModelState.IsValid)
             {
-                return View($"project-details/{details.Id}");
+                return BadRequest();
             }
 
             await _projectService.EditProject(details);
